registration and auth impl - refactoring
