+++ security/src/main/java/com/jlg/submatch/service/authentication/config/SecurityConfiguration.java
@@ -1,7 +1,6 @@
-package com.jlg.submatch.service.user.config;
+package com.jlg.submatch.service.authentication.config;
 
-import com.jlg.submatch.service.user.JwtAuthenticationFilter;
-import lombok.RequiredArgsConstructor;
+import com.jlg.submatch.service.authentication.JwtAuthenticationFilter;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.authentication.AuthenticationProvider;
